requirements.txt
matplotlib==2.2.3
opencv_python==4.1.2.30
torchvision==0.4.0a0
torchray==1.0.0.2
tqdm==4.45.0
pandas==0.23.3
scikit_image==0.15.0
numpy==1.14.3
torch==1.2.0
Pillow==7.1.2
scikit_learn==0.22.2.post1
skimage==0.0
